Remove Subscriptions from BaseFragment and BaseActivity
