+++ app/src/main/java/de/czyrux/store/ui/base/BaseFragment.java
@@ -11,12 +11,9 @@ import butterknife.ButterKnife;
 import butterknife.Unbinder;
 import io.reactivex.disposables.CompositeDisposable;
 import io.reactivex.disposables.Disposable;
-import rx.Subscription;
-import rx.subscriptions.CompositeSubscription;
 
 public abstract class BaseFragment extends Fragment {
 
-    private final CompositeSubscription subscriptions = new CompositeSubscription();
     private final CompositeDisposable disposables = new CompositeDisposable();
     private Unbinder unbinder;
 
@@ -39,12 +36,7 @@ public abstract class BaseFragment extends Fragment {
     @Override
     public void onStop() {
         super.onStop();
-        subscriptions.clear();
         disposables.dispose();
-    }
-
-    protected final void addSubscritiption(Subscription subscription) {
-        subscriptions.add(subscription);
     }
 
     protected final void addDisposable(Disposable disposable) {
