+++ app/src/main/java/de/czyrux/store/ui/catalog/CatalogFragment.java
@@ -19,7 +19,7 @@ import de.czyrux.store.core.domain.product.ProductResponse;
 import de.czyrux.store.core.domain.product.ProductService;
 import de.czyrux.store.inject.Injector;
 import de.czyrux.store.ui.base.BaseFragment;
-import de.czyrux.store.util.RxUtil2;
+import de.czyrux.store.util.RxUtil;
 import hu.akarnokd.rxjava.interop.RxJavaInterop;
 
 public class CatalogFragment extends BaseFragment implements CatalogListener {
@@ -74,8 +74,8 @@ public class CatalogFragment extends BaseFragment implements CatalogListener {
         showProgressBar();
 
         addDisposable(RxJavaInterop.toV2Observable(productService.getAllCatalog())
-                .compose(RxUtil2.applyStandardSchedulers())
-                .subscribe(this::onProductResponse, RxUtil2.emptyConsumer()));
+                .compose(RxUtil.applyStandardSchedulers())
+                .subscribe(this::onProductResponse, RxUtil.emptyConsumer()));
     }
 
     private void showProgressBar() {
@@ -112,8 +112,8 @@ public class CatalogFragment extends BaseFragment implements CatalogListener {
     public void onProductClicked(Product product) {
         CartProduct cartProduct = CartProductFactory.newCartProduct(product, 1);
         addDisposable(RxJavaInterop.toV2Observable(cartService.addProduct(cartProduct))
-                .compose(RxUtil2.applyStandardSchedulers())
-                .subscribeWith(RxUtil2.emptyObserver()));
+                .compose(RxUtil.applyStandardSchedulers())
+                .subscribeWith(RxUtil.emptyObserver()));
 
         Toast.makeText(getContext(), "Adding to cart..." + product.title, Toast.LENGTH_SHORT).show();
     }
