+++ app/src/main/java/de/czyrux/store/util/RxUtil.java
@@ -1,27 +1,26 @@
 package de.czyrux.store.util;
 
-import android.util.Log;
+import io.reactivex.ObservableTransformer;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.functions.Consumer;
+import io.reactivex.internal.functions.Functions;
+import io.reactivex.observers.DisposableObserver;
+import io.reactivex.schedulers.Schedulers;
 
-import rx.Observable;
-import rx.Observer;
-import rx.android.schedulers.AndroidSchedulers;
-import rx.functions.Action1;
-import rx.schedulers.Schedulers;
-
+/**
+ * Utility methods for RxJava 2
+ */
 public class RxUtil {
 
-    private static final Action1<Throwable> EMPTY_ERROR = throwable -> {
-    };
-
-    private static final Action1<Throwable> LOG_ERROR = throwable -> Log.e("Observable", Log.getStackTraceString(throwable));
-
-    private static final Observer EMPTY_OBSERVER = new Observer() {
-        @Override
-        public void onCompleted() {
-        }
+    private static final DisposableObserver EMPTY_OBSERVER = new DisposableObserver() {
 
         @Override
         public void onError(Throwable e) {
+        }
+
+        @Override
+        public void onComplete() {
+
         }
 
         @Override
@@ -29,23 +28,20 @@ public class RxUtil {
         }
     };
 
+
     private RxUtil() {
     }
 
-    public static Action1<Throwable> silentError() {
-        return EMPTY_ERROR;
-    }
-
-    public static Action1<Throwable> logError() {
-        return LOG_ERROR;
-    }
-
     @SuppressWarnings("unchecked")
-    public static <T> Observer<T> emptyObserver() {
+    public static <T> DisposableObserver<T> emptyObserver() {
         return EMPTY_OBSERVER;
     }
 
-    public static <T> Observable.Transformer<T, T> applyStandardSchedulers() {
+    public static Consumer<? super Throwable> emptyConsumer() {
+        return Functions.emptyConsumer();
+    }
+
+    public static <T> ObservableTransformer<T, T> applyStandardSchedulers() {
         return observable -> observable.subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread());
     }
